Move mysql jdbc into jdbc folder
