+++ agent/agent-plugins/jdbc/mysql8/src/main/java/org/bithon/agent/plugin/jdbc/mysql8/MySql8Plugin.java
@@ -1,0 +1,82 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.agent.plugin.jdbc.mysql8;
+
+import org.bithon.agent.instrumentation.aop.interceptor.descriptor.InterceptorDescriptor;
+import org.bithon.agent.instrumentation.aop.interceptor.matcher.Matchers;
+import org.bithon.agent.instrumentation.aop.interceptor.plugin.IPlugin;
+import org.bithon.agent.instrumentation.aop.interceptor.precondition.IInterceptorPrecondition;
+import org.bithon.shaded.net.bytebuddy.description.modifier.Visibility;
+
+import java.util.Arrays;
+import java.util.List;
+
+import static org.bithon.agent.instrumentation.aop.interceptor.descriptor.InterceptorDescriptorBuilder.forClass;
+
+/**
+ * @author frankchen
+ */
+public class MySql8Plugin implements IPlugin {
+    @Override
+    public IInterceptorPrecondition getPreconditions() {
+        return IInterceptorPrecondition.isClassDefined("com.mysql.cj.interceptors.QueryInterceptor");
+    }
+
+    @Override
+    public List<InterceptorDescriptor> getInterceptors() {
+
+        return Arrays.asList(
+
+            // PreparedStatement
+            forClass("com.mysql.cj.jdbc.ClientPreparedStatement")
+                .onMethod(Matchers.names("execute", "executeQuery", "executeUpdate"))
+                .andVisibility(Visibility.PUBLIC)
+                .andNoArgs()
+                .interceptedBy("org.bithon.agent.plugin.jdbc.mysql8.ClientPreparedStatement$Execute")
+                .build(),
+
+            // Statement
+            forClass("com.mysql.cj.jdbc.StatementImpl")
+                .onMethod(Matchers.names("execute", "executeQuery", "executeUpdate", "executeLargeUpdate"))
+                .andVisibility(Visibility.PUBLIC)
+                .andArgs(0, "java.lang.String")
+                .interceptedBy("org.bithon.agent.plugin.jdbc.mysql8.StatementImpl$Execute")
+
+                .onMethod(Matchers.names("executeBatch", "executeLargeBatch"))
+                .andVisibility(Visibility.PUBLIC)
+                .andNoArgs()
+                .interceptedBy("org.bithon.agent.plugin.jdbc.mysql8.StatementImpl$ExecuteBatch")
+
+                .build(),
+
+            // IO
+            forClass("com.mysql.cj.protocol.a.NativeProtocol")
+                .onMethod("sendCommand")
+                .andArgs("com.mysql.cj.protocol.Message", "boolean", "int")
+                .interceptedBy("org.bithon.agent.plugin.jdbc.mysql8.NativeProtocol$SendCommand")
+
+                /*
+                .onMethod("readAllResults")
+                .andArgs("int", "boolean", "com.mysql.cj.protocol.a.NativePacketPayload", "boolean",
+                         "com.mysql.cj.protocol.ColumnDefinition",
+                         "com.mysql.cj.protocol.ProtocolEntityFactory")
+                .interceptedBy("org.bithon.agent.plugin.jdbc.mysql8.NativeProtocolInterceptor")
+                */
+                .build()
+        );
+    }
+}
