--- agent/agent-plugins/mysql-8.x/pom.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0">
-
-  <parent>
-    <groupId>org.bithon.agent</groupId>
-    <artifactId>agent-all</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <relativePath>../../pom.xml</relativePath>
-  </parent>
-  <artifactId>agent-plugin-mysql-8.x</artifactId>
-  <modelVersion>4.0.0</modelVersion>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.bithon.agent</groupId>
-      <artifactId>agent-observability</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.bithon.shaded</groupId>
-      <artifactId>shaded-alibaba-druid</artifactId>
-      <version>1.0-SNAPSHOT</version>
-    </dependency>
-
-    <dependency>
-      <groupId>mysql</groupId>
-      <artifactId>mysql-connector-java</artifactId>
-      <version>8.0.30</version>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <finalName>${project.artifactId}</finalName>
-  </build>
-</project>
